feat: token 발급 api 분할, stock 배치 개수 출력
